refactor: move new BlockData CODEC from Particle to Block class
